+++ ReleaseNotes.md
@@ -4,6 +4,7 @@
 * **NEW FEATURE:** Added debug logging for `GetHttp` and `PostHttp` endpoints.
 * **NEW FEATURE:** String body of `PUT` endpoints now default to empty string.
 * **NEW FEATURE:** Added method `GetHttp`, synchronous version of `GetHttpAsync`.
+* **NEW FEATURE:** Added methods `GetCookies()` and `GetCookie(string)` to access rest client cookies.
 * **BUG FIX:** `PutHttp` endpoints now correctly call `PUT` instead of `POST`.
 
 ## v1.15.0